pyproject.toml
[project]
name = "flysearch"
version = "0.1.0"
description = "FlySearch: Exploring how vision-language models explore"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "gymnasium>=1.2.0",
    "matplotlib>=3.10.6",
    "numpy>=2.3.2",
    "openai>=1.105.0",
    "pandas>=2.3.2",
    "pillow>=11.3.0",
    "python-dotenv>=1.1.1",
    "scipy>=1.16.1",
    "seaborn>=0.13.2",
    "tqdm>=4.67.1",
    "typer>=0.17.3",
]

[project.optional-dependencies]
anthropic = [
    "anthropic>=0.66.0",
]
webgui = [
    "fastapi>=0.116.1",
    "uvicorn>=0.35.0"
]
gemini = [
    "google-genai>=1.49.0",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "mkdocs-material[imaging]>=9.6.20"
]
